feat: add `Board::display_with_promotions` method
This method effectively does the same thing as `Board::board_fen`.
However, instead of allocating a new `String` for each invocation, it
returns a `struct` that implements `Display` instead. In doing so, the
caller has the freedom to choose where to redirect the formatted output.

Hence, the `board_fen` method has been deprecated in favor of the
more generic `display_with_promotions` method. The `board_fen` method is
now implemented in terms of the latter for backwards-compatibility.