test: use `unwrap` instead of `?` operator
The `?` operator fails in `no_std` because the doc-tests implicitly cast
the error variant as a `Box<dyn std::error::Error>`. But since `no_std`
does not link with the `std` crate to begin with, the doc-tests will fail.

The solution is to `unwrap` instead. It is slightly more verbose, but it
gets the tests to pass.