Use circular instead of slice-deque
This probably performs worse, but it removes platform dependent code and
it removes all uses of unsafe (which makes me feel better).

This is for the moment a straight forward replacement with the minimal
changes necessary to make it compile, circular::Buffer has some methods
that could be used to tune when the buffer is shifted. We  might want to
take a look at optimizing the size of the buffer as well.